[Regen] Move buildFileTree to FileUtils
