+++ src/test/scala/temple/builder/project/ProjectBuilderTest.scala
@@ -3,10 +3,10 @@ package temple.builder.project
 import java.nio.file.Paths
 
 import org.scalatest.FlatSpec
-import temple.TestUtils
 import temple.detail.LanguageDetail.GoLanguageDetail
 import temple.generate.FileMatchers
 import temple.generate.FileSystem.File
+import temple.utils.FileUtils
 
 class ProjectBuilderTest extends FlatSpec with FileMatchers {
 
@@ -17,14 +17,14 @@ class ProjectBuilderTest extends FlatSpec with FileMatchers {
   behavior of "ProjectBuilder"
 
   it should "correctly create a simple project using postgres as the default" in {
-    val expected = TestUtils.buildFileMap(Paths.get("src/test/resources/project-builder-simple"))
+    val expected = FileUtils.buildFileMap(Paths.get("src/test/resources/project-builder-simple"))
     val actual =
       ProjectBuilder.build(ProjectBuilderTestData.simpleTemplefile, GoLanguageDetail("github.com/squat/and/dab"))
     projectFilesShouldMatch(actual, expected)
   }
 
   it should "use postgres when defined at the project level" in {
-    val expected = TestUtils.buildFileMap(Paths.get("src/test/resources/project-builder-simple"))
+    val expected = FileUtils.buildFileMap(Paths.get("src/test/resources/project-builder-simple"))
     val actual = ProjectBuilder
       .build(
         ProjectBuilderTestData.simpleTemplefilePostgresProject,
@@ -34,7 +34,7 @@ class ProjectBuilderTest extends FlatSpec with FileMatchers {
   }
 
   it should "use postgres when defined at the service level" in {
-    val expected = TestUtils.buildFileMap(Paths.get("src/test/resources/project-builder-simple"))
+    val expected = FileUtils.buildFileMap(Paths.get("src/test/resources/project-builder-simple"))
     val actual = ProjectBuilder
       .build(ProjectBuilderTestData.simpleTemplefilePostgresService, GoLanguageDetail("github.com/squat/and/dab"))
     projectFilesShouldMatch(actual, expected)
@@ -49,7 +49,7 @@ class ProjectBuilderTest extends FlatSpec with FileMatchers {
   }
 
   it should "correctly create a complex service with nested struct" in {
-    val expected = TestUtils.buildFileMap(Paths.get("src/test/resources/project-builder-complex"))
+    val expected = FileUtils.buildFileMap(Paths.get("src/test/resources/project-builder-complex"))
     val actual =
       ProjectBuilder.build(ProjectBuilderTestData.complexTemplefile, GoLanguageDetail("github.com/squat/and/dab"))
     projectFilesShouldMatch(actual, expected)
